backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
openai
transformers
accelerate
scikit-learn
python-dateutil
pydantic
google-generativeai
loguru
transformers>=4.41.1
datasets>=2.19.1
torch>=2.2.2        # cpu build
scikit-learn>=1.5.0 # still needed for metrics
evaluate
chart.js            # frontend CDN only
